pipeline/requirements.txt
openai>=1.30.0
supabase>=2.3.0
feedparser>=6.0.11
requests>=2.32.0
beautifulsoup4>=4.12.0
pydantic>=2.9.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0

loguru>=0.7.0
pyyaml>=6.0.2
